Find posts by a given date
We are not sure if here would be the right place/file to add this feature but @lele140686 and me came into this solution for a project where we created a simple API to check wordpress posts and we needed to get the posts for a given date. We would like to submit it for @ekino to review and merge it if this is fine.